Console: Move ConsoleBaseWidget from the Editor to it
